+++ deepracer/logs/__init__.py
@@ -1,4 +1,4 @@
 from .log_utils import ActionBreakdownUtils, AnalysisUtils, EvaluationUtils, NewRewardUtils, \
     PlottingUtils, SimulationLogsIO
-from .load_metrics import TrainingMetrics
+from .metrics import TrainingMetrics
 from .log import DeepRacerLog
